refactor: extract HTML report generation to separate file
